+++ packages/ckeditor5-ui/tests/labeledfield/utils.js
@@ -6,7 +6,8 @@
 import {
 	createLabeledInputText,
 	createLabeledInputNumber,
-	createLabeledDropdown
+	createLabeledDropdown,
+	createLabeledTextarea
 } from '../../src/labeledfield/utils';
 
 import LabeledFieldView from '../../src/labeledfield/labeledfieldview';
@@ -14,6 +15,7 @@ import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 import InputTextView from '../../src/inputtext/inputtextview';
 import InputNumberView from '../../src/inputnumber/inputnumberview';
 import DropdownView from '../../src/dropdown/dropdownview';
+import { TextareaView } from '@ckeditor/ckeditor5-ui';
 
 describe( 'LabeledFieldView utils', () => {
 	let locale;
@@ -172,6 +174,79 @@ describe( 'LabeledFieldView utils', () => {
 			labeledInput.errorText = 'some error';
 			labeledInput.fieldView.fire( 'input' );
 			expect( labeledInput.errorText ).to.be.null;
+		} );
+	} );
+
+	describe( 'createLabeledTextarea()', () => {
+		let labeledTextarea;
+
+		beforeEach( () => {
+			labeledTextarea = new LabeledFieldView( locale, createLabeledTextarea );
+		} );
+
+		afterEach( () => {
+			labeledTextarea.destroy();
+		} );
+
+		it( 'should create an TextareaView instance', () => {
+			expect( labeledTextarea.fieldView ).to.be.instanceOf( TextareaView );
+		} );
+
+		it( 'should pass the Locale to the textarea', () => {
+			expect( labeledTextarea.fieldView.locale ).to.equal( locale );
+		} );
+
+		it( 'should set textarea #id and #ariaDescribedById', () => {
+			labeledTextarea.render();
+
+			expect( labeledTextarea.fieldView.id ).to.equal( labeledTextarea.labelView.for );
+			expect( labeledTextarea.fieldView.ariaDescribedById ).to.equal( labeledTextarea.statusView.element.id );
+		} );
+
+		it( 'should bind textarea\'s #isReadOnly to labeledTextarea#isEnabled', () => {
+			labeledTextarea.isEnabled = true;
+			expect( labeledTextarea.fieldView.isReadOnly ).to.be.false;
+
+			labeledTextarea.isEnabled = false;
+			expect( labeledTextarea.fieldView.isReadOnly ).to.be.true;
+		} );
+
+		it( 'should bind textarea\'s #hasError to labeledTextarea#errorText', () => {
+			labeledTextarea.errorText = 'some error';
+			expect( labeledTextarea.fieldView.hasError ).to.be.true;
+
+			labeledTextarea.errorText = null;
+			expect( labeledTextarea.fieldView.hasError ).to.be.false;
+		} );
+
+		it( 'should bind labeledTextarea#isEmpty to input\'s #isEmpty', () => {
+			labeledTextarea.fieldView.isEmpty = true;
+			expect( labeledTextarea.isEmpty ).to.be.true;
+
+			labeledTextarea.fieldView.isEmpty = false;
+			expect( labeledTextarea.isEmpty ).to.be.false;
+		} );
+
+		it( 'should bind labeledTextarea#isFocused to textarea\'s #isFocused', () => {
+			labeledTextarea.fieldView.isFocused = true;
+			expect( labeledTextarea.isFocused ).to.be.true;
+
+			labeledTextarea.fieldView.isFocused = false;
+			expect( labeledTextarea.isFocused ).to.be.false;
+		} );
+
+		it( 'should bind labeledTextarea#placeholder to textarea\'s #placeholder', () => {
+			labeledTextarea.fieldView.placeholder = null;
+			expect( labeledTextarea.placeholder ).to.be.null;
+
+			labeledTextarea.fieldView.placeholder = 'foo';
+			expect( labeledTextarea.placeholder ).to.equal( 'foo' );
+		} );
+
+		it( 'should clean labeledTextarea#errorText upon textarea\'s DOM "update" event', () => {
+			labeledTextarea.errorText = 'some error';
+			labeledTextarea.fieldView.fire( 'input' );
+			expect( labeledTextarea.errorText ).to.be.null;
 		} );
 	} );
