+++ packages/ckeditor5-ui/theme/components/spinner/spinner.css
@@ -10,9 +10,6 @@
 .ck.ck-spinner-container {
 	display: block;
 	position: relative;
-	width: var(--ck-toolbar-spinner-size);
-	height: var(--ck-toolbar-spinner-size);
-	animation: 1.5s infinite rotate linear;
 }
 
 .ck.ck-spinner {
@@ -21,17 +18,6 @@
 	left: 0;
 	right: 0;
 	margin: 0 auto;
-	width: var(--ck-toolbar-spinner-size);
-	height: var(--ck-toolbar-spinner-size);
-	border-radius: 50%;
-	border: 2px solid var(--ck-color-text);
-	border-top-color: transparent;
 	transform: translateY(-50%);
 	z-index: 1;
-}
-
-@keyframes rotate {
-	to {
-		transform: rotate(360deg)
-	}
 }
